Add loading indicator.
Added state isLoading and logic to control when Loading indicator is shown. Shows when user clicks "New Joke" --> activates clickhandler which in turn runs the fetch function, where isLoading is set to false, data loads and is returned in a response. After the response the isLoading state is set to false so the joke is shown in render. Added Styling.

fixes issue: https://github.com/vino-dk/random-joke/issues/1